feat: no more uneeded metrics (preparation)
All rancher charts do not allow the manipulation of the metric_relabelings,
which allows keeping or droping metrics after them being scraped but right
before ingestion (when their __name__ is available). This blocks minimization
of the metrics collected in all subcharts of the monitoring stack.

To allow customization, the helper templates were all modified, to allow users to really define
their own metricRelabelings in their endpoints and also still keep the label replacements that
rancher did. As a positive aspect, the relabellings are done now to less metrics, so some RAM usage should be spared.

This commit includes the reduced metrics as a default value for the hardenedKubelet. If users need more metrics,
the relabelings can ofc be removed.

Signed-off-by: Bruno Bressi <bruno.bressi@telekom.de>