Fix/rancher performance (#34)
* chore: use release name for label

Signed-off-by: Bruno Bressi <bruno.bressi@telekom.de>

* chore: bumped to latest prom crds

* chore: removed uneeded instruction in make chart

The dependencies are already updated, no need to run the command again.

* fix: removed pinned images

The pinning of the images practically blocked the update of most
components.

* chore: removed unused value

Signed-off-by: Bruno Bressi <bruno.bressi@telekom.de>

* feat: added rancher-monitoring RBAC & smon

This commit adds the necessary templates to be able to parse the rancher
deployments metrics. The resources were copied from the official
rancher-monitoring chart.

The default behavior should be not to activate this monitoring, because
the chart is typically deployed in downstream clusters, which don't have
any rancher deployments.

* chore: rc0 for 1.0.0

Signed-off-by: Bruno Bressi <bruno.bressi@telekom.de>

* chore: rc1 for 1.0.0

Signed-off-by: Bruno Bressi <bruno.bressi@telekom.de>

---------

Signed-off-by: Bruno Bressi <bruno.bressi@telekom.de>