fix: added missing kubelet ingress policy
This policy was unfortunately forgotten when adding the hardened ingress policies for the
other pushprox pods. This blocked prometheus from scraping the kubelet of all other nodes except
the node on which the proxy itself was running, probably because of the standard np-default policy (which should
generally allow all incoming traffic from within the project to all pods of the project - anyway!).

Signed-off-by: Bruno Bressi <bruno.bressi@telekom.de>